split server objects into own file
